Move core.cpp/h to primitives.  Changed miner createblock back to CTranaction.
